moved definition
