+++ bindings/wasm/src/resolver/supported_document_types.rs
@@ -82,9 +82,6 @@ extern "C" {
   #[wasm_bindgen(typescript_type = "CoreDocument | IAsCoreDocument | undefined")]
   pub type OptionSupportedDocument;
 
-  #[wasm_bindgen(typescript_type = "Array<CoreDocument | IAsCoreDocument>")]
-  pub type ArrayIAsCoreDocument;
-
   #[wasm_bindgen(typescript_type = "Array<CoreDocument | IAsCoreDocument> | undefined")]
   pub type OptionArraySupportedDocument;
